fix: Don't call TerminateProcess with zero pid
After db174d0e it is possible to connect to a Starcraft2 instance
started by another entity (e.g. Ladder Manager). In this case the API
doesn't control the corresponding process of the Starcraft game client
but by some reason creates ProcessInfo structure setting pid 0 into it.

Later, when game ends, the API executes TerminateProcess on the
preserved process info, finds pid zero and tries to terminate it.
When it happens on Linux kill(0) is executed which finishes not only the
current sc2 instance, but every other process in the current process
group (the current bot process and Ladder Manager which started it).

This commit assumes that the API shouldn't pretend that it can control
a StarCraft instance launched by somebody else and does connection
without saving any process information.

Signed-off-by: Alexander Kurbatov <alkurbatov@virtuozzo.com>