fix: Avoid crash of replay observer on relaunch
During replays observing the API can decide to relaunch game client e.g.
in case of some errors. Such relaunch means that old StarCraft 2
instance is killed a new instance is launched, which leads to recreation
of Client::Control interface. As a result each replay observer should be
reinitialized with new control object.

However, the current reinitialization routine refreshes control for
the replay_control_imp_  member, but not for observer_action_imp_. Such
behaviour leads to a dangling pointer every usage of which cause Access
Violation and crash the observer application.

That was fixed.

Signed-off-by: Alexander Kurbatov <alkurbatov@virtuozzo.com>