feat: Use full screen for first player only, #21
It is turned out that only the first player (first participant) is
allowed to control the game (commands from the second player are
ignored). As a result, there is no sense in keeping support of the full
screen mode for the second player, although it works on OS X.

This patch forces the windowed mode for the second player in any case.

Signed-off-by: Alexander Kurbatov <alkurbatov@virtuozzo.com>