fix: Abort game on unexpected cmdline options
Previously, the API ignored unexpected cmdline options (only some words
were printed into the output) everywhere, except the ArgParser::Parse
method. As a result, the valuable code was not executed, API proceeded
further with partly initialized input parameters (i.e. ignoring part of
the provided data) and the game launch could fail e.g. because of
initData error. Such behavior was seen in the replay observer
used by AIArena and caused a lot of confusion.

Bailout as soon as possible if we can't parse the input args to avoid
such situations in the future.

Signed-off-by: Alexander Kurbatov <Alexander.Kurbatov@acronis.com>