fix: Make '-e, --executable' optional
API always tries to determine the location of the SC2 executable. After
it, the '-e, --executable' option is parsed and if it is omitted,
parsing immediately ends even if we have successfully detected
the executable previously. As a result, we can't run tests.

That was fixed by marking '-e, --executable' as optional.

Signed-off-by: Alexander Kurbatov <Alexander.Kurbatov@acronis.com>