build: msvc runtime library abstraction
Replaced MSVC configuration introduced in [0] with a CMake
MSVC runtime library abstraction. The original configuration
presumably relied on [1], and while there is no explicit citation
the block is identical and predates the commit by 5 years. Reliance on
the configuration was a consequence of CMake's limitations predating
version 3.15 and policy CMP0091 outlined in [2].

The option for building dynamically has also been removed. Linking to
submodules would error out when MSVC_RUNTIME did not default to static.

Refs:
    [0] 7286ea4000820a90bc5335b16e881ebec849365f
    [1] https://stackoverflow.com/questions/10113017/setting-the-msvc-runtime-in-cmake
    [2] https://cmake.org/cmake/help/latest/policy/CMP0091.html