build: cmake project restructure
The src directory contained several libraries, and the CMakeLists.txt
file for the directory consolidated build configuration for all of
the cpp-sc2 libraries.

In the restructure each library is built with a dedicated subdirectory
CMakeLists.txt file, isolating the build logic and providing specific
target include and linking definitions. Extraneous compile flags which
are either already set in the parent CMakeLists.txt - or were tested
as having no impact on the build are cleaned up.

A new interface library encompassing all of the cpp-sc2 linking and
include requirements was added to simplify the CMakeLists.txt file
configuration for bot executables.

An edge case of civetweb not building with the debug postfix was also
addressed to ease future work with precompiled libraries.