build: disable ipv6-parse
ipv6-parse was introduced as a submodule in [0], but no source files or
dependencies include headers from it. The sole source file which is
likely to have referenced it is the example project source file
bot_mp_ipv6.cc, which has never been added to the project nor enabled
(it has been commented out since it was merged in [1]).

Refs:
    [0] 7286ea4000820a90bc5335b16e881ebec849365f
    [1] https://github.com/cpp-sc2/cpp-sc2/commit/7286ea4000820a90bc5335b16e881ebec849365f#diff-940e5356fa1137e646ea135a6f68cacbfad4fe7124c3a69163468f588acf9283R43